+++ modules/civicrm/CRM/Core/BAO/EntityTag.php
@@ -2,9 +2,9 @@
 
 /*
  +--------------------------------------------------------------------+
- | CiviCRM version 3.3                                                |
+ | CiviCRM version 3.4                                                |
  +--------------------------------------------------------------------+
- | Copyright CiviCRM LLC (c) 2004-2010                                |
+ | Copyright CiviCRM LLC (c) 2004-2011                                |
  +--------------------------------------------------------------------+
  | This file is a part of CiviCRM.                                    |
  |                                                                    |
@@ -30,7 +30,7 @@
  * This class contains functions for managing Tag(tag) for a contact
  *
  * @package CRM
- * @copyright CiviCRM LLC (c) 2004-2010
+ * @copyright CiviCRM LLC (c) 2004-2011
  * $Id$
  *
  */
@@ -94,6 +94,13 @@ class CRM_Core_BAO_EntityTag extends CRM_Core_DAO_EntityTag
         // dont save the object if it already exists, CRM-1276
         if ( ! $entityTag->find( true ) ) {
             $entityTag->save( );
+			
+			//invoke post hook on entityTag
+        	require_once 'CRM/Utils/Hook.php';
+            // we are using this format to keep things consistent between the single and bulk operations
+            // so a bit different from other post hooks
+        	$object = array( 0 => array( 0 => $params['entity_id'] ), 1 => $params['entity_table'] );
+        	CRM_Utils_Hook::post( 'create', 'EntityTag', $params['tag_id'], $object );
         }
 
         return $entityTag;
@@ -127,8 +134,14 @@ class CRM_Core_BAO_EntityTag extends CRM_Core_DAO_EntityTag
     {
         $entityTag = new CRM_Core_BAO_EntityTag( );
         $entityTag->copyValues( $params );
-        $entityTag->delete( );
-        //return $entityTag;
+        if ( $entityTag->find( true ) ) {
+            $entityTag->delete( );
+		
+            //invoke post hook on entityTag
+            require_once 'CRM/Utils/Hook.php';
+            $object = array( 0 => array( 0 => $params['entity_id'] ), 1 => $params['entity_table'] );
+            CRM_Utils_Hook::post( 'delete', 'EntityTag', $params['tag_id'], $object );
+        }
     }
 
     /**
@@ -145,6 +158,8 @@ class CRM_Core_BAO_EntityTag extends CRM_Core_DAO_EntityTag
     static function addEntitiesToTag( &$entityIds, $tagId, $entityTable = 'civicrm_contact' ) {
         $numEntitiesAdded    = 0;
         $numEntitiesNotAdded = 0;
+		$entityIdsAdded      = array();
+		
         foreach ( $entityIds as $entityId ) {
             $tag = new CRM_Core_DAO_EntityTag( );
 
@@ -153,6 +168,7 @@ class CRM_Core_BAO_EntityTag extends CRM_Core_DAO_EntityTag
             $tag->entity_table  = $entityTable;
             if ( ! $tag->find( ) ) {
                 $tag->save( );
+				$entityIdsAdded[] = $entityId;
                 $numEntitiesAdded++;
             } else {
                 $numEntitiesNotAdded++;
@@ -161,7 +177,7 @@ class CRM_Core_BAO_EntityTag extends CRM_Core_DAO_EntityTag
 
         //invoke post hook on entityTag
         require_once 'CRM/Utils/Hook.php';
-        $object = array( $entityIds, $entityTable );
+        $object = array( $entityIdsAdded, $entityTable );
         CRM_Utils_Hook::post('create', 'EntityTag', $tagId, $object );
 
         // reset the group contact cache for all groups
@@ -187,6 +203,8 @@ class CRM_Core_BAO_EntityTag extends CRM_Core_DAO_EntityTag
     {
         $numEntitiesRemoved    = 0;
         $numEntitiesNotRemoved = 0;
+		$entityIdsRemoved      = array();
+		
         foreach ( $entityIds as $entityId ) {
             $tag = new CRM_Core_DAO_EntityTag( );
             
@@ -195,6 +213,7 @@ class CRM_Core_BAO_EntityTag extends CRM_Core_DAO_EntityTag
             $tag->entity_table = $entityTable;
             if (  $tag->find( ) ) {
                 $tag->delete( );
+				$entityIdsRemoved[] = $entityId;
                 $numEntitiesRemoved++;
             } else {
                 $numEntitiesNotRemoved++;
@@ -203,7 +222,7 @@ class CRM_Core_BAO_EntityTag extends CRM_Core_DAO_EntityTag
         
         //invoke post hook on entityTag
         require_once 'CRM/Utils/Hook.php';
-        $object = array( $entityIds, $entityTable );
+        $object = array( $entityIdsRemoved, $entityTable );
         CRM_Utils_Hook::post( 'delete', 'EntityTag', $tagId, $object );
         
         // reset the group contact cache for all groups
@@ -331,5 +350,56 @@ class CRM_Core_BAO_EntityTag extends CRM_Core_DAO_EntityTag
          
          return $entityTags;
     }  
+
+    /** 
+     * Function to merge two tags: tag B into tag A.
+     */
+    function mergeTags( $tagAId, $tagBId ) {
+        $queryParams = array( 1 => array($tagBId, 'Integer'),
+                              2 => array($tagAId, 'Integer') );
+
+        // re-compute used_for field
+        $query = "SELECT id, name, used_for FROM civicrm_tag WHERE id IN (%1, %2)";
+        $dao   = CRM_Core_DAO::executeQuery( $query, $queryParams );
+        $tags  = array( );
+        while( $dao->fetch( ) ) {
+            $label = ( $dao->id == $tagAId ) ? 'tagA' : 'tagB';
+            $tags[$label] = $dao->name;
+            $tags["{$label}_used_for"]  = $dao->used_for ? explode( ",", $dao->used_for ) : array( );
+        }
+        $usedFor = array_merge( $tags["tagA_used_for"], $tags["tagB_used_for"] );
+        $usedFor = implode( ',', array_unique($usedFor) );
+        $tags["tagB_used_for"] = explode( ",", $usedFor );
+
+        // get all merge queries together
+        $sqls   = array( 
+                        // 1. update entity tag entries
+                        "UPDATE civicrm_entity_tag SET tag_id = %1 WHERE tag_id = %2",
+                        // 2. update used_for info for tag B
+                        "UPDATE civicrm_tag SET used_for = '{$usedFor}' WHERE id = %1",
+                        // 3. remove tag A, if tag A is getting merged into B
+                        "DELETE FROM civicrm_tag WHERE id = %2",
+                        // 4. remove duplicate entity tag records
+                        "DELETE et1.* from civicrm_entity_tag et1 
+INNER JOIN ( SELECT * FROM civicrm_entity_tag 
+GROUP BY entity_table, entity_id, tag_id HAVING count(*) > 1 ) et2 ON et1.id = et2.id",
+                         );
+        $tables = array( 'civicrm_entity_tag', 'civicrm_tag' );
+
+        // Allow hook_civicrm_merge() to add SQL statements for the merge operation AND / OR 
+        // perform any other actions like logging
+        CRM_Utils_Hook::merge( 'sqls', $sqls, $tagAId, $tagBId, $tables );
+        
+        // call the SQL queries in one transaction
+        require_once 'CRM/Core/Transaction.php';
+        $transaction = new CRM_Core_Transaction( );
+        foreach ( $sqls as $sql ) {
+            CRM_Core_DAO::executeQuery( $sql, $queryParams, true, null, true );
+        }
+        $transaction->commit( );
+
+        $tags['status'] = true;
+        return $tags;
+    }
 }
